# vulkan: optimize flash attention split_k_reduce